bug fix to updateWeightTime, changed float to double to avoid machine truncation errors
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@7793 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/KernelConn.cpp
@@ -568,7 +568,7 @@ int KernelConn::updateWeights(int arbor_ID){
    return PV_BREAK;
 }
 
-float KernelConn::computeNewWeightUpdateTime(double time, double currentUpdateTime) {
+double KernelConn::computeNewWeightUpdateTime(double time, double currentUpdateTime) {
    // Is only called by KernelConn::updateState if plasticityFlag is true
    weightUpdateTime += weightUpdatePeriod;
    return weightUpdateTime;
